ncbi_tools/retrieval_package/requirements.txt
# NCBI Data Retriever Requirements
# Install these packages using: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0          # For making HTTP requests to NCBI API
PyYAML>=6.0               # For reading YAML configuration files

# File format support
pandas>=1.5.0             # For reading CSV and Excel files
openpyxl>=3.0.0           # For reading Excel (.xlsx) files
python-docx>=0.8.11       # For reading Word (.docx) files

# Additional dependencies that might be missing
lxml>=4.6.0               # Required by python-docx
et-xmlfile>=1.1.0         # Required by openpyxl
python-dateutil>=2.8.0    # Required by pandas
pytz>=2021.1              # Required by pandas
numpy>=1.21.0             # Required by pandas

# Development and testing (optional)
# pytest>=7.0.0           # For running tests
# black>=22.0.0            # For code formatting
# flake8>=5.0.0            # For code linting